refactor: rename event checklist modules for clarity
- event_checklist (resource status) → event_resources_checklist
- event_master_checklist (todo checklist) → event_checklist
Renames AMD modules, templates, PHP output classes, lang strings
and view file references consistently.
