+++ lang/de/bookit.php
@@ -177,17 +177,11 @@ $string['event_bookingstatus_3'] = 'Storniert';
 $string['event_bookingstatus_4'] = 'Abgelehnt';
 $string['event_bookingstatus_help'] = 'Erklärung der Buchungsstatus-Optionen.';
 $string['event_bookingstatus_list'] = 'Neu, In Bearbeitung, Bestätigt, Storniert, Abgelehnt';
-$string['event_checklist:booked_amount'] = 'Gebucht';
-$string['event_checklist:confirmed'] = 'bestätigt';
-$string['event_checklist:progress'] = 'Ressourcenstatus Fortschritt';
-$string['event_checklist_heading'] = 'Ressourcen-Checkliste für Veranstaltung: {$a}';
-$string['event_checklist_no_resources'] = 'Dieser Veranstaltung sind keine Ressourcen zugewiesen.';
-$string['event_checklist_title'] = 'Veranstaltungs-Ressourcen-Checkliste';
-$string['event_checklist_view:done'] = 'erledigt';
-$string['event_checklist_view:progress'] = 'Checklisten-Fortschritt';
-$string['event_checklist_view_heading'] = 'Checkliste für Veranstaltung: {$a}';
-$string['event_checklist_view_no_items'] = 'Für diese Veranstaltung sind keine Checklisten-Einträge verfügbar.';
-$string['event_checklist_view_title'] = 'Veranstaltungs-Checkliste';
+$string['event_checklist:done'] = 'erledigt';
+$string['event_checklist:progress'] = 'Checklisten-Fortschritt';
+$string['event_checklist_heading'] = 'Checkliste für Veranstaltung: {$a}';
+$string['event_checklist_no_items'] = 'Für diese Veranstaltung sind keine Checklisten-Einträge verfügbar.';
+$string['event_checklist_title'] = 'Veranstaltungs-Checkliste';
 $string['event_compensationfordisadvantages'] = 'Weitere Nachteilsausgleiche';
 $string['event_compensationfordisadvantages_help'] = 'Tragen Sie hier die bereits bekannten Informationen zu Studierenden mit Nachteilsausgleich ein.';
 $string['event_department'] = 'Institution';
@@ -212,6 +206,12 @@ $string['event_refcourseid'] = 'Prüfungskurs';
 $string['event_refcourseid_help'] = 'Auswahl des Prüfungskurses, der zu dieser Prüfung gehört.';
 $string['event_reserved'] = 'Gebucht';
 $string['event_resources'] = 'Ressourcen';
+$string['event_resources_checklist:booked_amount'] = 'Gebucht';
+$string['event_resources_checklist:confirmed'] = 'bestätigt';
+$string['event_resources_checklist:progress'] = 'Ressourcenstatus Fortschritt';
+$string['event_resources_checklist_heading'] = 'Ressourcen-Checkliste für Veranstaltung: {$a}';
+$string['event_resources_checklist_no_resources'] = 'Dieser Veranstaltung sind keine Ressourcen zugewiesen.';
+$string['event_resources_checklist_title'] = 'Veranstaltungs-Ressourcen-Checkliste';
 $string['event_resources_heading'] = 'Ressourcen für Veranstaltung: {$a}';
 $string['event_resources_title'] = 'Veranstaltungsressourcen';
 $string['event_room'] = 'Raum';
