+++ lang/en/bookit.php
@@ -184,17 +184,11 @@ $string['event_bookingstatus_3'] = 'Canceled';
 $string['event_bookingstatus_4'] = 'Rejected';
 $string['event_bookingstatus_help'] = 'Explanation of the booking status options.';
 $string['event_bookingstatus_list'] = 'New, In progress, Accepted, Canceled, Rejected';
-$string['event_checklist:booked_amount'] = 'Booked';
-$string['event_checklist:confirmed'] = 'confirmed';
-$string['event_checklist:progress'] = 'Resource Status Progress';
-$string['event_checklist_heading'] = 'Resource Checklist for Event: {$a}';
-$string['event_checklist_no_resources'] = 'No resources are assigned to this event.';
-$string['event_checklist_title'] = 'Event Resource Checklist';
-$string['event_checklist_view:done'] = 'done';
-$string['event_checklist_view:progress'] = 'Checklist Progress';
-$string['event_checklist_view_heading'] = 'Checklist for Event: {$a}';
-$string['event_checklist_view_no_items'] = 'No checklist items are available for this event.';
-$string['event_checklist_view_title'] = 'Event Checklist';
+$string['event_checklist:done'] = 'done';
+$string['event_checklist:progress'] = 'Checklist Progress';
+$string['event_checklist_heading'] = 'Checklist for Event: {$a}';
+$string['event_checklist_no_items'] = 'No checklist items are available for this event.';
+$string['event_checklist_title'] = 'Event Checklist';
 $string['event_compensationfordisadvantages'] = 'Other compensations of disadvantages';
 $string['event_compensationfordisadvantages_help'] = 'Enter here information on compensation of disadvantages.';
 $string['event_department'] = 'Department';
@@ -219,6 +213,12 @@ $string['event_refcourseid'] = 'Exam course';
 $string['event_refcourseid_help'] = 'Exam course associated with this exam';
 $string['event_reserved'] = 'Reserved';
 $string['event_resources'] = 'Resources';
+$string['event_resources_checklist:booked_amount'] = 'Booked';
+$string['event_resources_checklist:confirmed'] = 'confirmed';
+$string['event_resources_checklist:progress'] = 'Resource Status Progress';
+$string['event_resources_checklist_heading'] = 'Resource Checklist for Event: {$a}';
+$string['event_resources_checklist_no_resources'] = 'No resources are assigned to this event.';
+$string['event_resources_checklist_title'] = 'Event Resource Checklist';
 $string['event_resources_heading'] = 'Resources for Event: {$a}';
 $string['event_resources_title'] = 'Event Resources';
 $string['event_room'] = 'Room';
