+++ templates/event_resources_checklist_item_row.mustache
@@ -15,8 +15,8 @@
     along with Moodle.  If not, see <https://www.gnu.org/licenses/>.
 }}
 {{!
-    @template mod_bookit/event_checklist_item_row
-    Event checklist item row template
+    @template mod_bookit/event_resources_checklist_item_row
+    Event resources checklist item row template
 
     Context variables:
     * id           - bookit_event_resource record ID
@@ -48,8 +48,8 @@
         "isrejected": false
     }
 }}
-<tr class="event-checklist-item-row"
-    data-region="event-checklist-item-row"
+<tr class="event-resources-checklist-item-row"
+    data-region="event-resources-checklist-item-row"
     data-itemid="{{id}}"
     data-item-resourceid="{{resourceid}}"
     data-item-status="{{status}}">
